requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-dotenv==1.0.0
python-multipart==0.0.6
bcrypt==4.0.1
pyjwt==2.8.0
cryptography==41.0.7
pydantic==2.5.0
email-validator==2.1.0
pyyaml==6.0.1
